+++ pi_cpp_pthreadGlobalState.cpp
@@ -1,7 +1,7 @@
 /*
  *  A C program to calculate Pi using quadrature as a threads-based algorithm.
  *
- *  Copyright © 2009-10 Russel Winder
+ *  Copyright © 2009--2011 Russel Winder
  */
 
 #include <iostream>
@@ -9,10 +9,10 @@
 #include <pthread.h>
 #include "microsecondTime.h"
 
-const long n = 1000000000l ;
+const int n = 1000000000 ;
 const double delta = 1.0 / n ;
 
-long sliceSize  ;
+int sliceSize  ;
 
 double sum ;
 pthread_mutex_t sumMutex ;
@@ -29,10 +29,10 @@ typedef int integerFromVoidStar ;
 #endif
 
 void * partialSum ( void *const arg  ) {
-  const long start = 1 + ( (integerFromVoidStar) arg ) * sliceSize ;
-  const long end = ( ( (integerFromVoidStar) arg ) + 1 ) * sliceSize ;
+  const int start = 1 + ( (integerFromVoidStar) arg ) * sliceSize ;
+  const int end = ( ( (integerFromVoidStar) arg ) + 1 ) * sliceSize ;
   double localSum = 0.0 ;
-  for ( long i = start ; i <= end ; ++i ) {
+  for ( int i = start ; i <= end ; ++i ) {
     const double x = ( i - 0.5 ) * delta ;
     localSum += 1.0 / ( 1.0 + x * x ) ;
   }
